+++ swanlab/data/callback_cloud.py
@@ -9,8 +9,6 @@ r"""
 """
 from .run.callback import MetricInfo, ColumnInfo
 from swanlab.cloud import UploadType
-from urllib.parse import quote
-from swanlab.error import ApiError
 from swanlab.api.upload.model import ColumnModel, ScalarModel, MediaModel
 from swanlab.api import LoginInfo, create_http, terminal_login
 from swanlab.api.upload import upload_logs
