+++ swanlab/data/callback_local.py
@@ -7,12 +7,10 @@ r"""
 @Description:
     基本回调函数注册表，此时不考虑云端情况
 """
-from typing import Callable
 from swanlab.log import swanlog
 from swanlab.utils.font import FONT
 from swanlab.data.run.main import get_run, SwanLabRunState
 from swanlab.data.run.callback import SwanLabRunCallback, MetricInfo, RuntimeInfo
-from swanlab.data.system import get_system_info, get_requirements
 from swanlab.env import ROOT
 from datetime import datetime
 import traceback
@@ -107,28 +105,19 @@ class LocalRunCallback(SwanLabRunCallback):
             return swanlog.debug("SwanLab Runtime has been cleaned manually.")
         self._train_finish_print()
         # 如果正在运行
-        run.finish() if run.is_running else swanlog.debug("Duplicate finish, ignore it.")
+        run.finish() if run.running else swanlog.debug("Duplicate finish, ignore it.")
 
     def on_init(self, proj_name: str, workspace: str, logdir: str = None):
         self._init_logdir(logdir)
 
-    def before_init_experiment(
-        self,
-        run_id: str,
-        exp_name: str,
-        description: str,
-        num: int,
-        suffix: str,
-        setter: Callable[[str, str, str, str], None],
-    ):
-        requirements_path = self.settings.requirements_path
-        metadata_path = self.settings.metadata_path
-        # 将实验依赖存入 requirements.txt
-        with open(requirements_path, "w") as f:
-            f.write(get_requirements())
-        # 将实验环境(硬件信息、git信息等等)存入 swanlab-metadata.json
-        with open(metadata_path, "w") as f:
-            json.dump(get_system_info(self.settings.version, self.settings.swanlog_dir), f)
+    def on_runtime_info_update(self, r: RuntimeInfo):
+        # 更新运行时信息
+        if r.requirements is not None:
+            r.requirements.write(self.settings.files_dir)
+        if r.metadata is not None:
+            r.metadata.write(self.settings.files_dir)
+        if r.config is not None:
+            r.config.write(self.settings.files_dir)
 
     def on_run(self):
         swanlog.install(self.settings.console_dir)
