+++ swanlab/data/run/system/bin/apple_gpu_stats
